Complete the channel created within a category.

+++ app/entry.py
@@ -22,19 +22,19 @@ async def on_voice_state_update(member, before, after):
     if before.channel == after.channel:
         return
 
-    if before.channel.id in tmp_channels.keys():
+    if len(tmp_channels) > 0 and before.channel.id in tmp_channels.keys():
         if len(before.channel.members) == 0:
             await before.channel.delete()
             del tmp_channels[before.channel.id]
             print('delete channel: {}'.format(before.channel.id))
 
-    if after.channel.id in channels.keys():
+    if len(channels) > 0 and after.channel.id in channels.keys():
         type = channels[after.channel.id]
-        guild = after.channel.guild
 
         limit = channel_types[type]
         channel_name = "{0}_{1}".format(type.upper(), datetime.datetime.now().strftime('%Y%m%d%H%M%S'))
-        tmp_channel = await guild.create_voice_channel(channel_name, user_limit=limit)
+        category = after.channel.category
+        tmp_channel = await category.create_voice_channel(channel_name, user_limit=limit)
         print("create channel: {}".format(tmp_channel.id))
         tmp_channels[tmp_channel.id] = type
         await member.move_to(tmp_channel)
